@stratusjs/idx moving sync variables to one location
